# Use ChainExists to reduce memory footprint