@tnqn 2021-07-23
> Is the github.com/coreos/go-iptables upgrade opportunistic or is it required to use `ChainExists` (to be clear I am fine with the upgrade even if it's opportunistic...)
> 
> I think this should also be backported...

Thanks for the quick review. 

go-iptables added `ChainExists` in 0.5.0 https://github.com/coreos/go-iptables/commit/b15012c6af6a0b27ffb80954ac32d44eee91c937, I bumped up to latest version which includes a fix for timeout.

Sure, will backport it.